# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type GullDropOffer = typeof GullDropOffer[keyof typeof GullDropOffer];


export const GullDropOffer = {
  messy: 'messy',
  patron: 'patron',
  founder: 'founder',
} as const;

export interface GullDropCheckoutInput {
  origin: string;
  offer: GullDropOffer;
  /**
     * @minLength 4
     * @maxLength 48
     * @pattern ^[A-Za-z0-9_-]+$
     */
  deviceId: string;
  /**
     * @maxLength 32
     * @pattern ^[a-z0-9_-]*$
     */
  source?: string;
  /**
     * @maxLength 32
     * @pattern ^[a-z0-9_-]*$
     */
  creator?: string;
}

export interface GullDropCheckout {
  url: string;
}

export interface GullDropFulfillment {
  paid: boolean;
  offer: GullDropOffer;
  /** @minimum 0 */
  amount: number;
  source?: string;
  creator?: string;
}

export type GullDropLeaderboardMode = typeof GullDropLeaderboardMode[keyof typeof GullDropLeaderboardMode];


export const GullDropLeaderboardMode = {
  classic: 'classic',
  armed: 'armed',
} as const;

export interface GullDropLeaderboardEntry {
  id: number;
  handle: string;
  score: number;
  combo: number;
  mode: GullDropLeaderboardMode;
  /** @nullable */
  weapon?: string | null;
  createdAt: string;
}

export type GullDropScoreInputWeapon = typeof GullDropScoreInputWeapon[keyof typeof GullDropScoreInputWeapon];


export const GullDropScoreInputWeapon = {
  street: 'street',
  messy: 'messy',
  patron: 'patron',
  fizz: 'fizz',
  founder: 'founder',
} as const;

export interface GullDropScoreInput {
  /**
     * @minLength 6
     * @maxLength 16
     * @pattern ^[A-Z0-9-]+$
     */
  handle: string;
  /**
     * @minimum 1
     * @maximum 3000000
     */
  score: number;
  /**
     * @minimum 0
     * @maximum 500
     */
  combo: number;
  mode: GullDropLeaderboardMode;
  weapon?: GullDropScoreInputWeapon;
  /**
     * @minimum 5000
     * @maximum 95000
     */
  durationMs: number;
}

export type GetGullDropFulfillmentParams = {
/**
 * @pattern ^cs_
 */
sessionId: string;
/**
 * @minLength 4
 * @maxLength 48
 * @pattern ^[A-Za-z0-9_-]+$
 */
deviceId: string;
};

export type ListGullDropLeaderboardParams = {
mode: GullDropLeaderboardMode;
};