# lib/api-zod/src/generated/types/gullDropFulfillment.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GullDropOffer } from './gullDropOffer';

export interface GullDropFulfillment {
  paid: boolean;
  offer: GullDropOffer;
  /** @minimum 0 */
  amount: number;
  source?: string;
  creator?: string;
}