# lib/api-zod/src/generated/types/gullDropLeaderboardEntry.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GullDropLeaderboardMode } from './gullDropLeaderboardMode';

export interface GullDropLeaderboardEntry {
  id: number;
  handle: string;
  score: number;
  combo: number;
  mode: GullDropLeaderboardMode;
  /** @nullable */
  weapon?: string | null;
  createdAt: Date;
}