# lib/api-zod/src/generated/types/gullDropLeaderboardMode.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type GullDropLeaderboardMode = typeof GullDropLeaderboardMode[keyof typeof GullDropLeaderboardMode];


export const GullDropLeaderboardMode = {
  classic: 'classic',
  armed: 'armed',
} as const;